simplify matrix by using matrix_commons from tegonal-gh-commons
